move singleton into utils file
