+++ rdkit_mcp/utils.py
@@ -1,4 +1,8 @@
 from typing import Callable
+import logging
+
+
+logger = logging.getLogger(__name__)
 
 
 def is_rdkit_tool(func: Callable) -> bool:
@@ -7,3 +11,21 @@ def is_rdkit_tool(func: Callable) -> bool:
     while hasattr(func, "__wrapped__"):
         func = func.__wrapped__
     return hasattr(func, "_is_rdkit_tool")
+
+
+def singleton(cls):
+    """Add as a decorator to a class to make it a singleton."""
+    instances = {}
+
+    def get_instance(*args, **kwargs):
+        if cls not in instances:
+            logger.debug("Creating singleton instance of %s with args: %s, kwargs: %s", cls.__name__, args, kwargs)
+            instances[cls] = cls(*args, **kwargs)
+        elif kwargs and instances[cls] is not None:
+            # If kwargs are provided, create a new instance
+            # This fixes issue where AppSettings can be instantiated before yaml is loaded,
+            # and then we can never get the correct settings.
+            logger.debug("Overwriting singleton instance of %s with kwargs: %s", cls.__name__, kwargs)
+            instances[cls] = cls(*args, **kwargs)
+        return instances[cls]
+    return get_instance
